Refactorings for simplification (split responsibilities in different classes). Still some refactoring to do (some methods are too complex).
Code separated in new tag (Installer) in Reflectivity.

Added API + code for objects and classes. The user is now able to manipulate links on objects and classes without browsing the ast. 

Work in Progress: commited for the day (too much changes at once) and everything is not covered by tests yet.